requirements.txt
arrow==0.10.0
attrs==16.3.0
cffi==1.9.1
click==6.6
constantly==15.1.0
coverage==4.2
cryptography==1.6
duo-client==3.0
elasticsearch==2.4.0
elasticsearch-dsl==2.2.0
enum34==1.1.6
Flask==0.11.1
Flask-Script==2.0.5
funcsigs==1.0.2
google-api-python-client==1.5.5
httplib2==0.9.2
idna==2.1
incremental==16.10.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
klein==15.3.1
Logbook==1.0.0
MarkupSafe==0.23
mock==2.0.0
ndg-httpsclient==0.4.2
netaddr==0.7.18
nose==1.3.7
ntlm-auth==1.0.2
oauth2client==3.0.0
oauthlib==2.0.1
ordereddict==1.1
pbr==1.10.0
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyJWT==1.4.2
pymssql==2.1.3
pyOpenSSL==16.2.0
pytest==3.0.5
pytest-cov==2.4.0
pytest-logbook==1.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests[security]==2.12.3
requests-ntlm==1.0.0
requests-oauthlib==0.7.0
retrying==1.3.3
rsa==3.4.2
service-identity==16.0.0
six==1.10.0
stevedore==1.19.1
treq==15.1.0
Twisted==16.6.0
txretry==1.0.1
git+https://github.com/wrapp/txwebretry.git@df5014019cf3a5d501bc755b9ce46509c1b31351#egg=txwebretry
uritemplate==3.0.0
urllib3==1.19.1
validate-email==1.3
Werkzeug==0.11.11
zope.interface==4.3.2
